Check if policies.json already exists

+++ brave-bullshitinator-linux-install.sh
@@ -5,7 +5,9 @@ echo "Brave debullshitinator is made by Mules Gaming. Find it here: https://gith
 sudo mkdir /etc/brave/policies/managed -p
 # Download file
 wget https://raw.githubusercontent.com/samtowers/brave-debloatinator/main/policies.json
-sudo rm /etc/brave/policies/managed/policies.json
+if [ -f /etc/brave/policies/managed/policies.json ]; then
+    sudo rm /etc/brave/policies/managed/policies.json
+fi
 sudo cp policies.json /etc/brave/policies/managed
 rm policies.json
 # End
